Update GAS prebuilts from Binutils 2.38
This is now built on QEMU system running Debian 10.11, which is the latest
oldstable release at time of compilation.

Debian 10 ships with Glibc 2.28, however prebuilts should be able to run on
system running at least Glibc 2.14 or newer, although this isn't always going
to be guaranteed.

These binaries are tuned for Skylake hosts.

Signed-off-by: Albert I <kras@raphielgang.org>